Move DictDataset from test utils to API
