+++ tfep/io/dataset/__init__.py
@@ -1,2 +1,3 @@
 from tfep.io.dataset.traj import TrajectoryDataset, TrajectorySubset, get_subsampled_indices
 from tfep.io.dataset.merged import MergedDataset
+from tfep.io.dataset.dict import DictDataset
